+++ src/main/java/com/telepathicgrunt/worldblender/entities/WBEntities.java
@@ -1,6 +1,8 @@
 package com.telepathicgrunt.worldblender.entities;
 
 import com.telepathicgrunt.worldblender.WorldBlender;
+import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
+import net.minecraft.entity.EntityDimensions;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.SpawnGroup;
 import net.minecraft.util.Identifier;
@@ -8,8 +10,7 @@ import net.minecraft.util.registry.Registry;
 
 public class WBEntities {
 
-    public static final EntityType<ItemClearingEntity> ITEM_CLEARING_ENTITY = EntityType.Builder.<ItemClearingEntity>
-                    create(ItemClearingEntity::new, SpawnGroup.MISC).setDimensions(0, 0).maxTrackingRange(0).build("item_clearing_entity");
+    public static final EntityType<ItemClearingEntity> ITEM_CLEARING_ENTITY = FabricEntityTypeBuilder.<ItemClearingEntity>create(SpawnGroup.MISC, ItemClearingEntity::new).dimensions(EntityDimensions.fixed(0, 0)).trackRangeChunks(0).build();
 
     public static void registerEntities() {
         Registry.register(Registry.ENTITY_TYPE, new Identifier(WorldBlender.MODID, "item_clearing_entity"), ITEM_CLEARING_ENTITY);